+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 1.19.0 (Unreleased)
+
+ENHANCEMENTS: 
+* resource/spotinst_elastigroup_aws: added support for `max_capacity_per_scale` under `scaling_target_policy`
+
 ## 1.18.0 (June 24, 2020)
 
 BUG FIXES: